[SEQNG-333] Display of Nod and Shuffle Step

+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/semanticui/elements/progress/Progress.scala
@@ -17,8 +17,8 @@ import web.client.facades.semanticui.SemanticUIProgress._
   */
 object Progress {
   final case class Props(label:       String,
-                         total:       Long,
-                         value:       Long,
+                         total:       Int,
+                         value:       Int,
                          indicating:  Boolean = false,
                          progress:    Boolean = false,
                          color:       Option[String] = None,
